Port screenshot handling to gtk3
WARNING: the port was done by applying verbatim the porting rules from
GdkRegion -> cairo_region_t etc. It should have instead be done with
more care and understanding of the code, but since the previous patches
left the app not building, I figured it was better to just commit this.