screenshot: Rewrite event handling in area selection
The previous code was handling X events directly. This breaks when GTK
decides to use XI2 and no more regular events arrive.
Instead we now use the window we create anyway to catch the events. At
the window, they'll arrive properly preprocessed by GTK in the way we
expect them.
As a bonus, this code does not rely on X11 anymore, so we can stop
including gdkx.h.