Don't use window title in default filenames
Now that we have the timestamp it isn't needed to help make the filename
unique. It makes the filename really really long and makes the window
snaps inconsistent from the area and full screen ones. This makes it a
bit harder to edit name name if necessary in the dialog.

https://bugzilla.gnome.org/show_bug.cgi?id=565907