screenshot: compensate invisible window borders using frame extents
In order to grab a screenshot of a window we currently peek at the WM
decorated X window size and coordinates, and cut a pixbuf from the root
window using those dimensions.

This unfortunately breaks with invisible window borders, as the WM
window is actually bigger than expected, resulting in an unwanted frame
around the image. Mutter now stores the real dimensions of the window
decorated frame in _NET_FRAME_EXTENTS, so we can just use that
information to get the actual region to cut.

Mutter also stopped setting the XShape bounding box on rounded corners,
as they're now properly antialiased. Our code that takes the XShape
rectangles and makes the area outside of their bounding box transparent
also needs to be updated to handle the case where the WM decorated window
size is different from the frame extents.

Tested with GNOME Shell/Mutter and Metacity.

https://bugzilla.gnome.org/show_bug.cgi?id=659885