screenshot: don't apply some gsettings for headless mode
When we're not in interactive mode, we don't want to follow gsettings
preferences such as delay or including the window border.
Since changing options in interactive mode saves them to gsettings, we
would end up in a situation where e.g. setting a delay there would have
made that delay the default for key press invocation too.