screenshot: add an application menu for the interactive window
When started in interactive mode, the tool should behave as a real
GNOME app, and use an application menu.
This also allows us to remove the Help button from there, and move it to
the application menu.
Finally, the toplevel window is now a regular GtkApplicationWindow (not
a GtkDialog), so it gained a real "Close" button we can use instead of
the "Cancel" one we had before.

https://bugzilla.gnome.org/show_bug.cgi?id=674947