Fix launching from gnome-shell
Instead of starting in headless mode, start in the interactive mode and
bring up the UI when we're being DBus activated.

To make this possible, this patch separates the GSettings loading and
command line parsing, so we would only have to parse the command line
when started on the command line, and skip this in the DBus activation
case.

https://bugzilla.gnome.org/show_bug.cgi?id=735375