Drop intltool and clean up the build
We can use upstream gettext for everything we ship, these days.

While switching away from intltool I realised that the appdata XML we
ship is *ancient* and full of deprecated tags — to the point that
gettext's own ITS translation rules were unable to work with it; the
only reason intltool works is because it does not do any validation.