Move the screenshot sound effects to gnome-shell
Companion patches:
* gnome-shell https://gitlab.gnome.org/GNOME/gnome-shell/merge_requests/211
* g-s-d https://gitlab.gnome.org/GNOME/gnome-settings-daemon/merge_requests/44

Since it can take some time to compress a PNG image,
at very high resolutions there was a noticeable delay
before the audiovisual feedback on screenshots.

To keep from complicating the interface for the shell screenshot
service, moving the sound effects into the service (which already
did the visual effects) so it can trigger them after capture and
before compression.

Note this removes the canberra dependency for gnome-screenshot.