Remove support for removing window border
All Wayland applications are using client-side decorations. Even on X11,
most GNOME applications are using them, so this option doesn't have any
effects in majority of cases. Since !7, it's not exposed in GUI either.

To simplify the code, just remove it.

Deprecate --include-border and --remove-border CLI options, print a
warning when they are in use. Update the manpage.

Remove the corresponding gsettings key.