application: Drop support for non-unique mode in headless instance
gnome-shell started allowing only applications that own certain names
on the session bus to interact with the screenshot interface. While the
name used by gnome-screenshot is on the allowed list, non-unique
instances that are used for the headless mode do not own that name. Fix
this by also making the headless instance unique so it will own the
name. This however also means that there can only ever be a single
instance of gnome-screenshot, regardless of headless or interactive.

Fixes: https://gitlab.gnome.org/GNOME/gnome-screenshot/-/issues/179